split off electronics.py
