+++ fssh/__main__.py
@@ -22,9 +22,6 @@ from __future__ import print_function, division
 
 import numpy as np
 
-import math as m
-import multiprocessing as mp
-import collections
 import pickle
 
 import sys
