+++ fssh/surface.py
@@ -20,7 +20,6 @@
 
 from __future__ import print_function
 
-import sys
 import argparse
 import numpy as np
 import fssh
