+++ fssh/trajectory.py
@@ -22,10 +22,6 @@ from __future__ import print_function, division
 
 import numpy as np
 
-import math as m
-import multiprocessing as mp
-import collections
-
 from fssh import TraceManager
 
 #####################################################################################
